+++ tokio/src/macros/cfg.rs
@@ -353,3 +353,13 @@ macro_rules! cfg_uds {
         )*
     }
 }
+
+macro_rules! cfg_unstable {
+    ($($item:item)*) => {
+        $(
+            #[cfg(tokio_unstable)]
+            #[cfg_attr(docsrs, doc(cfg(tokio_unstable)))]
+            $item
+        )*
+    }
+}
