+++ tokio/src/sync/mod.rs
@@ -425,10 +425,10 @@ cfg_sync! {
 
     pub mod broadcast;
 
-    #[cfg(tokio_unstable)]
-    mod cancellation_token;
-    #[cfg(tokio_unstable)]
-    pub use cancellation_token::{CancellationToken, WaitForCancellationFuture};
+    cfg_unstable! {
+        mod cancellation_token;
+        pub use cancellation_token::{CancellationToken, WaitForCancellationFuture};
+    }
 
     pub mod mpsc;
